Refactor for use images from frontend folder
