+++ src/main/java/com/softserve/teachua/controller/PdfGenerationController.java
@@ -6,7 +6,7 @@ import com.softserve.teachua.dto.club.ClubResponse;
 import com.softserve.teachua.dto.search.SearchClubProfile;
 import com.softserve.teachua.service.CenterService;
 import com.softserve.teachua.service.ClubService;
-import com.softserve.teachua.service.documentreport.ReportGenerationService;
+import com.softserve.teachua.documentreport.ReportGenerationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
