+++ src/main/java/com/softserve/teachua/documentreport/ReportGenerationService.java
@@ -1,4 +1,4 @@
-package com.softserve.teachua.service.documentreport;
+package com.softserve.teachua.documentreport;
 
 public interface ReportGenerationService<T> {
     byte[] getPdfOutput(T t);
