+++ src/main/java/com/softserve/teachua/documentreport/impl/ResultSearchReportGenerationServiceImpl.java
@@ -1,7 +1,7 @@
-package com.softserve.teachua.service.documentreport.impl;
+package com.softserve.teachua.documentreport.impl;
 
+import com.softserve.teachua.documentreport.ReportGenerationService;
 import com.softserve.teachua.dto.club.ClubResponse;
-import com.softserve.teachua.service.documentreport.ReportGenerationService;
 import net.sf.jasperreports.engine.JRDataSource;
 import net.sf.jasperreports.engine.JRException;
 import net.sf.jasperreports.engine.JasperCompileManager;
